==> Scripts/11_RunSimulationStudy.R <==
###################################################################################################
### This script is used to run the simulation study found in the main body of the manuscript along 
### with Appendix B.2 - Part Three 
###
### L29 - L53 load the required libraries, assign the correct paths, and source in custom functions
###
### L55 - L99 contain the parameters settings that need to be selected before running 
### to_vary on L87-L97 contains the settings that change across each set of samples 
### (see methods of the manuscript)
###  
### L101 - L111 run and save the results of the simulation study (NOTE SLOW!)
###
### Any theoretically reasonable settings should work. However, make sure your chosen correlations 
### are within the bounds defined in Appendix A.2 (based on phi.f/phi.m or p.f/p.m) of the manuscript 
### otherwise the probabilities from the joint distribution of survival/recapture wont within [0,1]
### 
### NOTE: The output from each mark###.out file generated by program MARK will paste all of its 
### output into the console. When running a large study (several model parameters with thousands
### of datasets), these outputs can cause a huge slowdown. If the user has a Linux Machine or an
### Ubuntu terminal available via windows 10, calling this script in the terminal with 
###
### > Rscript 11_RunSimulationStudy.R
###
### is much faster than using a standard IDE like Rstudio, in my experience. If you find a way
### to stop these messages from appearing please let me know. I've tried source-sink and suppress() 
### type functions in R to no avail. 
####################################################################################################

###################################
# Set up libraries and directories#
###################################
## Libraries
libs <- c("tidyverse", "parallel","RMark","gridExtra")
## If package is not installed then do so
for (pkg in libs) {
  if (!pkg %in% rownames(installed.packages())) {
    install.packages(pkg)
  }
}
## Attach our libraries
lapply(libs, require, character.only = TRUE)
## Inline object concatenation (like python)
`%+%` <- function(a, b) paste0(a, b)
#Set Working Directory and reference paths
path2dat <- getwd() %+% "/Data/"
path2out <- getwd() %+% "/Output/"
path2scripts <- getwd() %+% "/Scripts/"

#####################################
# Call in code from relevant scripts#
#####################################
source(path2scripts %+% "01_Fn_SimPropModel.R")
source(path2scripts %+% "02_Fn_ModelCode.R")

#####################################
# Generate Mated CJS Simulation Data#
#####################################
# Baseline Parameters
n <- 200 ##Sample Size 
k <- 4 ##Sampling Occasions
delta <- rep(1,k) #Breeding Probability
phi.f <- rep(0.7,k) #Marginal Female Survival from j to j+1
phi.m <- rep(0.7,k) #Marginal Male Survival from j to j+1
phi <- rep(0.7,k) #Marginal survival for any animal
gamma <- rep(0,k) #Correlation between males and female survival
p.f <- rep(0.8,k) #Marginal Female Recapture at j
p.m <- rep(0.8,k) #Marginal Male Recapture at j
p <- rep(0.8,k) #Marginal recapture for any animal
rho <- rep(0,k) #Correlation between Female and Male Recapture
prob.female <- 0.5 #Proportion of females in the population

#Parameters
parameter_list <- list(n = n,
                       k = k,
                       delta = delta,
                       phi.f = phi.f,
                       phi.m = phi.m,
                       gamma = gamma,
                       p.f = p.f,
                       p.m = p.m,
                       rho = rho,
                       prob.female = prob.female,
                       phi = phi,
                       p = p)

#################
#Grid Simulation# 
#################
# Baseline Parameters that do not vary
to_vary <- list("phi.f" = seq(0.7,0.7,by=0.1),
                "phi.m"= seq(0.7,0.7,by=0.1),
                "p.f" = seq(0.8,0.8,by=0.1),
                "p.m" = seq(0.8,0.8,by=0.1),
                "phi" = seq(0.7,0.7,by=0.1),
                "p" = seq(0.8,0.8,by=0.1),
			        	"n" = c(200)) 
# Add parameters that do vary 
to_vary$gamma <- seq(-0.4, 1.0, 0.1) #Survival correlation 
to_vary$rho <- seq(-0.25, 1.0, 0.25)

# Split jobs into batches so as not to overload RAM 
batches <- generate_grid(parameter_list,to_vary)[[2]] %>% nrow()
#Simulate Grid (CHANGE Batch Save = T to save results along with batches)
mark_grid <- simulate_grid(parameter_list,to_vary,
                           model_choice = "CJS_dgr",
                           iterations = 10,
                           batches = batches,
                           model_groups=c("S","N"),
                           batch_save = F)
#Save Results (If only want to keep final result set batch_save = F in above function)
saveRDS(mark_grid,file=path2out  %+% Sys.Date() %+% "TEST-OUTPUT.rds")
